+++ nope/app/register/default.php
@@ -1,0 +1,6 @@
+<?php
+
+// Register/unregister anything
+
+#\Nope::unregisterTextFormat('html');
+\Nope::setDefaultTextFormat('markdown');