+++ nope/lib/routes/view.php
@@ -8,7 +8,7 @@ $app->group(NOPE_ADMIN_ROUTE . '/view', function() {
 
   $this->get('/{view:[a-zA-Z0-9-_\/]+}{ext:\.html}', function ($request, $response, $args) {
     if($args['view'] === 'app') {
-      $data['menuItems'] = \Nope::getConfig('nope.admin.menu');
+      $data['menuItems'] = \Nope::getMenuItems();
     } elseif (S::startsWith($args['view'], 'setting/detail/')) {
       $p = explode('/', $args['view']);
       $settingKey = $p[count($p)-1];